Admin finalizado y uso de caché

+++ app/Http/Controllers/Admin/PostController.php
@@ -8,25 +8,27 @@ use App\Models\Post;
 use App\Models\Tag;
 use Illuminate\Http\Request;
 use App\Http\Requests\PostRequest;
+use Illuminate\Support\Facades\Cache;
 use Illuminate\Support\Facades\Storage;
 
 class PostController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
+
+    public function __construct()
+    {
+        $this->middleware('can:admin.post.index')->only('index');
+        $this->middleware('can:admin.post.create')->only('create', 'store');
+        $this->middleware('can:admin.post.edit')->only('edit', 'update');
+        $this->middleware('can:admin.post.destroy')->only('destroy');
+    }
+
+
     public function index()
     {
         return view('admin.posts.index');
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
+
     public function create()
     {
         $categories = Category::pluck('name', 'id');
@@ -34,12 +36,7 @@ class PostController extends Controller
         return view('admin.posts.create', compact('categories', 'tags'));
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
+
     public function store(PostRequest $request)
     {
         $post = Post::create($request->all());
@@ -51,6 +48,8 @@ class PostController extends Controller
             ]);
         }
 
+        Cache::flush();
+
         if($request->tags){
             $post->tags()->attach($request->tags);
         }
@@ -58,39 +57,19 @@ class PostController extends Controller
         return redirect()->route('admin.posts.index')->with('info', 'Post creado con éxito');
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function show(Post $post)
-    {
-        return view('admin.posts.show', compact('post'));
-    }
 
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
     public function edit(Post $post)
     {
+        $this->authorize('author', $post);
         $categories = Category::pluck('name', 'id');
         $tags = Tag::all();
         return view('admin.posts.edit', compact('post', 'categories', 'tags'));
     }
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
+
     public function update(PostRequest $request, Post $post)
     {
+        $this->authorize('author', $post);
         $post->update($request->all());
         if ($request->file('file')) {
             $url = Storage::put('posts',$request->file('file'));
@@ -111,18 +90,17 @@ class PostController extends Controller
             $post->tags()->sync($request->tags);
         }
 
+        Cache::flush();
+
         return redirect()->route('admin.posts.index')->with('info', 'Post actualizado con éxto');
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
+
     public function destroy(Post $post)
     {
+        $this->authorize('author', $post);
         $post->delete();
+        Cache::flush();
         return redirect()->route('admin.posts.index')->with('info', 'Post Eliminado con éxito');
     }
 }
